Refactor history views to use offline storage meal type
